+++ saleor/graphql/account/mutations/staff.py
@@ -8,9 +8,9 @@ from graphql_jwt.exceptions import PermissionDenied
 from ....account import events as account_events, models, utils
 from ....account.error_codes import AccountErrorCode
 from ....account.thumbnails import create_user_avatar_thumbnails
+from ....account.utils import get_random_avatar
 from ....checkout import AddressType
 from ....core.permissions import get_permissions
-from ....core.utils import get_random_avatar
 from ....core.utils.url import validate_storefront_url
 from ....dashboard.emails import send_set_password_email_with_url
 from ....dashboard.staff.utils import remove_staff_member
